requirements.txt
Django==1.9.2
Pillow==3.1.1
beautifulsoup4==4.4.1
django-admin-sortable==2.0.12
django-debug-toolbar==1.4
django-livereload==1.5
psycopg2==2.6.1
sqlparse==0.1.18
